requirements.txt
flask
torch
torchvision
pillow
fpdf
tqdm
sqlite3 